+++ collector-bitmovin-player/src/test/java/com/bitmovin/analytics/bitmovin/player/BitmovinSdkAdapterTest.kt
@@ -63,6 +63,7 @@ class BitmovinSdkAdapterTest {
                     mockk(),
                     mockk(),
                     mockk(),
+                    mockk(),
                 ),
                 recordPrivateCalls = true,
             )