+++ collector-exoplayer/src/test/java/com/bitmovin/analytics/exoplayer/ExoPlayerAdapterTest.kt
@@ -40,6 +40,7 @@ class ExoPlayerAdapterTest {
                     mockk(),
                     mockk(),
                     mockk(),
+                    mockk(),
                 ),
                 recordPrivateCalls = true,
             )