+++ collector-media3-exoplayer/src/test/java/com/bitmovin/analytics/media3/exoplayer/Media3ExoPlayerAdapterTest.kt
@@ -40,6 +40,7 @@ class Media3ExoPlayerAdapterTest {
                     mockk(),
                     mockk(),
                     mockk(),
+                    mockk(),
                 ),
                 recordPrivateCalls = true,
             )